% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babygrowth.R
\docType{data}
\name{babygrowth}
\alias{babygrowth}
\title{Sample data}
\format{
An object of class \code{data.frame} with 12 rows and 4 columns.
}
\usage{
babygrowth
}
\description{
Imaginary babygrowth data
}
\examples{
babygrowth

}
\keyword{datasets}
