% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clean_NOAA}
\alias{clean_NOAA}
\title{The NOAA Significant Earthquake Database
A dataset containing location, strenght, and descriptive information for 5901 significant earthquake events worldwide since 2500 BCE}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5901 rows and 48 columns.}
\source{
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
}
\usage{
clean_NOAA
}
\description{
The NOAA Significant Earthquake Database
A dataset containing location, strenght, and descriptive information for 5901 significant earthquake events worldwide since 2500 BCE
}
\keyword{datasets}
