% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\docType{data}
\name{earthquakes}
\alias{earthquakes}
\title{Earthquakes data}
\format{
A \code{tibble} with 6385
observations and 39 columns
\describe{   # Documenting some columns
\item{Year}{Year when earthquake occurred}
\item{Mo}{Month when earthquake occurred}
\item{Dy}{Day when earthquake occurred}
\item{Mag}{Magnitude of the earthquake}
\item{Longitude}{Longitude of location where the earthquake occurred}
\item{Latitude}{Latitude of location where the earthquake occurred}
}
}
\source{
From US NOAA's website
\url{https://www.ngdc.noaa.gov/hazel/view/hazards/earthquake}
}
\usage{
earthquakes
}
\description{
Earthquakes data
}
\examples{
data("earthquakes")

head(earthquakes)

earthquakes
}
\keyword{datasets}
