% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean the NOAA dataset}
\usage{
eq_clean_data(raw_data)
}
\arguments{
\item{raw_data}{Data frame from \emph{U.S. National Oceanographic and Atmospheric
Administration \strong{(NOAA)}}.
Below the download link \url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}}
}
\value{
A \emph{clean dataframe} without nans and clean columns to plot in the next steps.
If the dataframe is not the one corresponding to NOAA, it will return an error
when not finding the columns corresponding to it.
}
\description{
This function is in charge of creating a new column, \code{DATE} based on the \code{YEAR},
\code{MONTH} and \code{DAY} columns.
It also converts the \code{LATITUDE} and \code{LONGITUDE} columns to a numeric class.
Finally clean the \code{LOCATION_NAME} column to be able to plot later
with the relevant annotations.
}
\examples{
\dontrun{
data <- readr::read_delim('path_to_data', quote = "", delim = "\t")
data <- eq_clean_data(data)
}

}
