% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{eq_clean_location}
\alias{eq_clean_location}
\title{Helper for \code{eq_clean_data} function. Clean the column \code{LOCATION_NAME}.}
\usage{
eq_clean_location(raw_data)
}
\arguments{
\item{raw_data}{Data frame from \emph{U.S. National Oceanographic and Atmospheric
Administration \strong{(NOAA)}}.
Below the download link \url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}}
}
\value{
A \emph{clean dataframe} with the \code{LOCATION_NAME} column clean.
If the dataframe is not the one corresponding to NOAA, it will return an error
when not finding the columns corresponding to it.
}
\description{
This function covers the \code{LOCATION_NAME} column, extracts the country and
leaves the corresponding \code{LOCATION} visible.
}
\examples{
\dontrun{
data <- eq_clean_location(data)
}

}
