% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_tools.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Generates a label for a leaflet map}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{A filtered data frame containing cleaned NOAA earthquake data.}
}
\value{
A character vector with labels.
}
\description{
This function generates a label for the \code{leaflet} map based on location
name, magnitude and casualties from NOAA earthquake data.
}
