% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_location_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Modify rows or columns in a data frame}
\usage{
eq_location_clean(x, year_min = 1, year_max = 2050, countries = NULL)
}
\arguments{
\item{x}{data frame with \verb{Location Name} column to be cleaned.}

\item{year_min}{numeric to set minimum for years in DATE column
Optional argument with default value 1.}

\item{year_max}{numeric to set maximum for years in DATE column
Optional argument with default value 2050.}

\item{countries}{character vector containing a list of countries
used to filter the output. Optional argument
with default value NULL.}
}
\value{
data frame with cleaned COUNTRY and LOCATION_NAME columns
}
\description{
Create COUNTRY and Location_NAME column in a
data frame with a Location Name column as in NOAA's earthquakes
data frame. Clean Location Name column by splitting it
into COUNTRY and LOCATION_NAME columns where COUNTRY
contains countries names in upper case and LOCATION_NAME
contains locations names in title case. It also trims out
white space from the new names. The optional arguments:
year_min, year_max, countries can be used to filter the cleaned
for some range of years and some countries.
}
\examples{
# Clean earthquakes data frame
eq_location_clean(earthquakes)

# Clean earthquakes data and filter data for specific countries
eq_location_clean(earthquakes, countries = "ALASKA")
eq_location_clean(earthquakes, countries = c("PERU", "FRANCE", "ITALY"))

# Clean earthquakes data and filter data for a minimum or maximum year
eq_location_clean(earthquakes, year_min = 1970)
eq_location_clean(earthquakes, year_max = 1975)
eq_location_clean(earthquakes, year_min = 1970, year_max = 2000)

# Clean earthquakes data and filter data for specific countries
# and years range
eq_location_clean(earthquakes, 1970, 2000, c("PERU", "FRANCE", "ITALY"))

}
