% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Create interactive map with popup annotation}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{data frame with LONGITUDE, LATITUDE and
Mag columns}

\item{annot_col}{string used for popup annotation}
}
\value{
list or map
}
\description{
This function uses functions from \verb{leaflet package}
to create an interactive map with popup annotation.
The column to be used for annotation should be a
vector character or a string supplied through the
\code{annot_col} argument. Each row of the \code{data} supplied
is represented on the map with a \code{circle}  at LONGITUDE
and LATITUDE on that row. The radius of the circle is
proportional to the Mag on that row. The annotation is
displayed by clicking on the circle.
}
\examples{
library(magrittr)
library(leaflet)

df = data.frame(
Year = 2001:2005, Mo = 1:5, Dy = 11:15,
Longitude = 21:25, Latitude = 31:35,
`Location Name` = paste(LETTERS[1:5],
 LETTERS[1:5], sep = ":"),
 Mag = 3, Deaths = 0,
 check.names = FALSE
)

eq_location_clean(df) \%>\%
eq_map(annot_col = "DATE")

}
