% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Create Time Series Plots Of Earthquakes}
\usage{
geom_timeline(
  data = NULL,
  mapping = NULL,
  stat = "timeline",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{data}{data to be displayed in this layer}

\item{mapping}{set of aesthetics}

\item{stat}{statistical transformation to use on data}

\item{position}{position function to use}

\item{show.legend}{logical or NA.}

\item{inherit.aes}{logical to indicate if this geom function
should inherit aesthetics from the call to ggplot
or any other top geom.}

\item{na.rm}{logical to indicate how NAs should be removed:
with or without warning}

\item{...}{other parameters to be passed to layer()}
}
\value{
list or map
}
\description{
Layer function for ggproto GeomTimeline class
This function plots time series by adding a layer of
a point on horizontal line for each row in the plot
data supplied.
}
\examples{

# Example 1
lcd = eq_location_clean(earthquakes, countries = "ALASKA")
ggplot2::ggplot(lcd, ggplot2::aes(x = DATE)) +
geom_timeline()

# Example 2
lcd = eq_location_clean(earthquakes, 2000, 2020, c("CHINA","TURKEY"))
ggplot2::ggplot(
  lcd,
  ggplot2::aes(x = DATE, y = COUNTRY, colour = Deaths, size = Mag)) +
  geom_timeline()

# Example 3
df = data.frame(
DATE = c(1,2,3,1,2),
COUNTRY = c("C","A","C","B","B")
)

ggplot2::ggplot(df, ggplot2::aes(x = DATE, y = COUNTRY)) +
 geom_timeline(xMin = 2,xMax = 5)

# Example 4
data = babygrowth
ggplot2::ggplot(data, ggplot2::aes(x = DATE, y = 1, colour = height,
                                  size = weight)) +
 geom_timeline(alpha = .5)

# Example 5
data = babygrowth
ggplot2::ggplot(data, ggplot2::aes(x = DATE, y = AGE, colour = height,
                                  size = weight)) +
 geom_timeline()

# Example 6
data = my_EQ_clean
ggplot2::ggplot(data, ggplot2::aes(x = DATE, y = COUNTRY, colour = Deaths,
                                   size = Mag)) +
 geom_timeline()

# Example 7
data = my_EQ_clean
ggplot2::ggplot(data, ggplot2::aes(x = DATE, colour = Deaths, size = Mag)) +
  geom_timeline()

}
