% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{GeomTimeline
ggproto object of class Geom. Builds grobs for the layer function
geom_timeline()}
\arguments{
\item{x}{required aesthetic to map dates of occurrence of earthquakes}

\item{y}{string, required aesthetic to map countries of occurrence of earthquakes}

\item{color, }{default aesthetic to map number of deaths caused by earthquakes}

\item{alpha, }{default aesthetic for transparency}

\item{size}{default aesthetic to define size of points}

\item{shape}{default aesthetic for shape of points used.}

\item{linewidth, }{default aesthetic for horizontal lines}

\item{line_col, }{default aesthetic for color of horizontal lines}
}
\description{
GeomTimeline
ggproto object of class Geom. Builds grobs for the layer function
geom_timeline()
}
\keyword{internal}
