% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{my_draw_key_circle}
\alias{my_draw_key_circle}
\title{Allow geom_timeline to draw a useful legend; this function is completely internal to geom_timeline, and is not of use elsewhere}
\usage{
my_draw_key_circle(data, params, size)
}
\arguments{
\item{data}{The dataframe passed to the drawing key}

\item{params}{The params passed from the geom}

\item{size}{The size of the legend key}
}
\description{
Allow geom_timeline to draw a useful legend; this function is completely internal to geom_timeline, and is not of use elsewhere
}
\examples{
# no examples appropriate, this function is completely internal to geom_timeline
}
