% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_EQ.R
\docType{data}
\name{my_EQ}
\alias{my_EQ}
\title{Imaginary earthquakes data}
\format{
An object of class \code{data.frame} with 5 rows and 8 columns.
}
\usage{
my_EQ
}
\description{
Imaginary earthquakes data
}
\examples{
my_EQ

my_EQ_clean

}
\keyword{datasets}
