% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_timeline.R
\name{stat_timeline}
\alias{stat_timeline}
\title{Add layer: Time Series Plots}
\usage{
stat_timeline(
  mapping = NULL,
  data = NULL,
  geom = "timeline",
  position = "identity",
  xMin = as.Date("0001-01-01"),
  xMax = as.Date("2050-12-31"),
  show.legend = NA,
  inherit.aes = TRUE,
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{mapping, }{set of aesthetics}

\item{data, }{data used in statistical transformation}

\item{geom}{geom layer function to call}

\item{position}{where to plot each group of graphical element}

\item{xMin}{date parameter used to subset data}

\item{xMax}{date parameter used to subset data}

\item{show.legend}{logical or NA. TRUE always display legend,
FALSE does not. NA display legend only if suuplied.}

\item{inherit.aes}{logical to indicate whether to inherit aes
from ggplot or other top geoms}

\item{na.rm}{logical parameter to indicate how NAs data should be removed}

\item{...}{other arguments to be passed}
}
\description{
Compute layers to be added when mapping variables
in data transformed by StatTimeline to aesthetics
to create plots
}
\examples{
library(ggplot2)
lcd = eq_location_clean(
earthquakes, countries = c("PERU", "CHILE")
)
ggplot2::ggplot(lcd, ggplot2::aes(x = DATE, y = COUNTRY)) +
stat_timeline()

}
