% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_timeline_label.R
\name{stat_timeline_label}
\alias{stat_timeline_label}
\title{Add annotation layer}
\usage{
stat_timeline_label(
  mapping = NULL,
  data = NULL,
  geom = "timeline_label",
  position = "identity",
  show.legend = FALSE,
  inherit.aes = FALSE,
  n_max = NA,
  sgm_len = 0.1,
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{mapping}{set of aesthetics.}

\item{data}{to be displayed in this layer}

\item{geom}{other layer to use with this funcion}

\item{position}{position function}

\item{show.legend}{logical or NA. TRUE will always display
a legend. FALSE will not. NA will display
legend only when it is supplied.}

\item{inherit.aes}{Logical to indicate if this geom should
inherit aesthetics from the call to
ggplot or any other top geom}

\item{n_max}{numeric to set how many points to annotate}

\item{sgm_len}{numeric to set length of segment}

\item{na.rm}{Logical to indicate how NAs should be removed.
with warning or not.}

\item{...}{other arguments to be passed onto layer}
}
\description{
This geom function add annotation layer to a plot.
It constructs a segment line and add text annotation
from specified  column in the data
}
\examples{
data = babygrowth
ggplot2::ggplot(
 data, ggplot2::aes(x = DATE, y = AGE, colour = height,
                   size = weight)
) +
ggplot2::geom_point() +
stat_timeline_label(
ggplot2::aes(x = DATE, y = AGE, label = DATE))

}
