% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_timeline.R
\docType{data}
\name{StatTimeline}
\alias{StatTimeline}
\title{StatTimeline}
\arguments{
\item{x}{vector of type date, required aesthetic to
map dates of occurrence of earthquakes}

\item{y}{string, optional aesthetic}

\item{xMin}{date parameter used to subset data}

\item{xMax}{date parameter used to subset data}

\item{alpha}{transparency chanel for points.}

\item{colour}{default aesthetic for color points}

\item{size}{default aesthetic  for size of points}

\item{grp_aes}{default aesthetic to define grouping
structure in a data frame.}
}
\description{
ggproto object of Stat class to define aesthetics
and for statitical transformation
}
\keyword{internal}
