% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_timeline_label.R
\docType{data}
\name{StatTimelineLabel}
\alias{StatTimelineLabel}
\title{StatTimelineLabel}
\arguments{
\item{x}{vector of type date, required aesthetic to
map dates of occurrence of earthquakes}

\item{y}{string, optional aesthetic}

\item{label}{string used for annotation
required aesthetic.}

\item{by}{numeric optional aesthetic used to arrange data}

\item{data}{data frame to use in statistical transformation}

\item{n_max}{numeric to subset data before annotation}

\item{sgm_len}{numeric parameter to set length of segment
used in annotation}
}
\description{
ggproto object of Stat class to define aesthetics
and for statitical transformation
}
\keyword{internal}
