% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_timeline.R
\name{theme_timeline}
\alias{theme_timeline}
\title{Theme Timeline}
\usage{
theme_timeline(...)
}
\arguments{
\item{...}{any arguments to be passed to theme_classic}
}
\value{
\verb{theme element}.
}
\description{
Modify theme_classic() by removing the y_axis, its thicks
and title. It also sets the legend position at "bottom"
}
\examples{
library(ggplot2)
lcd = eq_location_clean(
earthquakes, countries = c("PERU", "CHILE")
)
ggplot2::ggplot(lcd, ggplot2::aes(x = DATE, y = COUNTRY)) +
stat_timeline() +
theme_timeline()

}
