% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USA_STATES.R
\docType{data}
\name{USA_STATES}
\alias{USA_STATES}
\title{USA STATES}
\format{
An object of class \code{character} of length 50.
}
\usage{
USA_STATES
}
\description{
A character vector of 50 US States
}
\keyword{datasets}
