% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Modify rows or columns in earthquakes data frame}
\usage{
eq_clean_data(x)
}
\arguments{
\item{x}{data frame with columns : Year, Mo, Dy, Longitude,
Latitude ...}
}
\value{
data frame
}
\description{
Clean data frame with specific columns as
in earthquakes data frame from NOAA including: \code{Year} ,
\code{Mo}, \code{Dy}, \code{Longitude}, \code{Latitude}. Unite columns: Year,
Mo and Dy to create a date class column DATE. It coerces
columns: Longitude and Latitude to numeric class columns:
LONGITUDE, LATITUDE. NAs are removed under columns: Year,
Mo, Dy used to form the new DATE column. Only Year > 0
are considered to ensure compatibility with lubridate package.
}
