% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_create_label.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create HTML text for popup annotation}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{data frame used to create the string}
}
\value{
character vector or string
}
\description{
Create string to be used for popup annotation on a map.
It requires columns: LOCATION_NAME, Total Deaths
and Mag from your data frame
}
\examples{
# Annotation text to use in popup
lcd = eq_location_clean(earthquakes, 2000, 2020, "JAPAN")
eq_create_label(lcd)

}
