% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_tools.R
\name{eq_map}
\alias{eq_map}
\title{Plot a leaflet map of earthquakes}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{A filtered data frame containing cleaned NOAA earthquake data.}

\item{annot_col}{A character. The name of the column in the data that should
be used as descriptor.}
}
\value{
A leaflet map with earthquakes and annotations. It will return an error
when not finding the right data frame or the \code{annot_col} is not of the
indicated type.
}
\description{
This function at takes an argument data containing the filtered data frame with
earthquakes to visualize and creates a \code{leaflet} map of selected earthquakes.
}
\examples{
\dontrun{
library(dplyr)

data <- readr::read_delim('path_to_data', quote = "", delim = "\t")

data \%>\%
   eq_clean_data() \%>\%
   dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
   dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
   eq_map(annot_col = "popup_text")
}

}
