% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Create Time Series Plots Of Earthquakes}
\usage{
geom_timeline(
  data = NULL,
  mapping = NULL,
  stat = "timeline",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{data}{data to be displayed in this layer}

\item{mapping}{set of aesthetics}

\item{stat}{statistical transformation to use on data}

\item{position}{position function to use}

\item{show.legend}{logical or NA.}

\item{inherit.aes}{logical to indicate if this geom function
should inherit aesthetics from the call to ggplot
or any other top geom.}

\item{na.rm}{logical to indicate how NAs should be removed:
with or without warning}

\item{...}{other parameters to be passed to layer()}
}
\value{
list or map
}
\description{
Layer function for ggproto GeomTimeline class
This function plots time series by adding a layer of
a point on horizontal line for each row in the plot
data supplied.
}
\examples{
\dontrun{
# Example 1
lcd = eq_location_clean(earthquakes, countries = "ALASKA")
ggplot2::ggplot(lcd, ggplot2::aes(x = DATE)) +
geom_timeline()

# Example 2
lcd = eq_location_clean(earthquakes, 2000, 2020,
c("CHINA","TURKEY"))
ggplot2::ggplot(
 lcd,
 ggplot2::aes(x = DATE, y = COUNTRY, colour = Deaths,
              size = Mag, group = 1:nrow(lcd))) +
 geom_timeline() + theme_timeline()

# Example 3
df = data.frame(
DATE = c(1,2,3,1,2),
COUNTRY = c("C","A","C","B","B")
)

ggplot2::ggplot(df, ggplot2::aes(x = DATE, y = COUNTRY)) +
 geom_timeline(xMin=1,xMax=3)

# Example 4
dat = data.frame(
DATE = as.Date(c("2010-10-23", "2010-12-24", "2010-5-13",
                "2010-5-27","2010-01-13","2010-7-18",
                "2010-03-16","2010-01-13","2010-7-18",
age = c(1,2,2,1,3,3,2,3,1,4,4,3),
height = c(1,2,3,1,3,3,4,3,4,5,7,5),
weight = c(2,1,3,2,2,3,4,2,3,6,6,5)
)

ggplot2::ggplot(dat, ggplot2::aes(x = DATE, y = 1, colour = height,
                                 size = weight, group = 1:nrow(dat))) +
 geom_timeline(alpha = .5) + theme_timeline()
}

}
