% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Geom for plotting a time line of earthquakes}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  ...,
  na.rm = FALSE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}
}
\description{
This function return a \strong{ggplot layer} that representing
representing a time line of earthquakes.
The size of the points is relatative to the earthquakes's magnitude, and
the color is related to the total number of deaths.
}
\details{
Aesthetics:
\code{geom_timeline} understands the following aesthetics
(required in bold):
\itemize{
\item \strong{x}: (Date) of earthquakes
\item y: (factr) stratification. If present multiple time lines
will be plotted for each level of the factor
(e.g. country).
\item colour: of the points
\item size: of the points
\item alpha: trasparency for the points
\item fill
\item linetype
\item linesize
\item fontsize
\item stroke
}
}
\examples{
\dontrun{
library(dplyr)

data <- readr::read_delim('path_to_data', quote = "", delim = "\t")

data \%>\%
   capstoner::eq_clean_data() \%>\%
   dplyr::filter(filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000)
   ) \%>\%
   ggplot2::ggplot(aes(
      x = date,
      y = country,
      size = eq_primary,
      colour = log(total_deaths)
      )) +
   geom_timeline()
}
}
