% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Add annotation to time series plots}
\usage{
geom_timeline_label(
  data = NULL,
  mapping = NULL,
  stat = "timeline_label",
  position = "identity",
  inherit.aes = FALSE,
  show.legend = FALSE,
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{data}{data frame to display in this layer}

\item{mapping}{set of aesthetics}

\item{stat}{statistical transformation to perform on data}

\item{position}{position function}

\item{inherit.aes}{logical to indicate if this geom function
should inherit aesthetics from the call to ggplot
or any other top geom.}

\item{show.legend}{logical or NA.}

\item{na.rm}{logical to indicate how NAs should be removed:
with or without warning}

\item{...}{other parameters to be passed to layer()}
}
\value{
list or map
}
\description{
This function bulds annotation layer by adding segment
and text label for each row of the plot data
}
\examples{
data = babygrowth

ggplot2::ggplot(
 data, ggplot2::aes(x = DATE, y = AGE, colour = height,
                   size = weight)
) +
ggplot2::geom_point() +
geom_timeline_label(
data, ggplot2::aes(x = DATE, y = AGE, label = DATE))

}
