% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{GeomTimelineLabel}
\arguments{
\item{x}{required aesthetic to map dates of occurrence of earthquakes}

\item{y}{string, required aesthetic to map countries of
occurrence of earthquakes}

\item{xend}{description}

\item{yend}{description}

\item{label}{description}

\item{by}{description}

\item{colour}{default aesthetic to map number of deaths
caused by earthquakes}

\item{angle}{description}

\item{alpha, }{default aesthetic for transparency}

\item{size}{default aesthetic to define size of points}

\item{hjust}{default aesthetic for shape of points used.}

\item{vjust}{description}

\item{linewidth, }{default aesthetic for vertical segments.}

\item{linetype}{description.}

\item{linecolour}{default aesthetic for color of horizontal lines.}

\item{lineaplha}{colours alpha channel for vertical segments.}
}
\description{
ggproto object of class Geom. Builds annotation grobs
for the layer function geom_timeline_label()
}
\keyword{internal}
